+++ packages/local_auth/local_auth/test/local_auth_test.dart
@@ -37,7 +37,6 @@ void main() {
         const AndroidAuthMessages(),
         const WindowsAuthMessages(),
       ],
-      options: const AuthenticationOptions(),
     )).called(1);
   });