+++ packages/webview_flutter/webview_flutter_android/android/src/test/java/io/flutter/plugins/webviewflutter/DownloadListenerTest.java
@@ -13,6 +13,7 @@ import static org.mockito.Mockito.verify;
 import android.webkit.DownloadListener;
 import io.flutter.plugins.webviewflutter.DownloadListenerHostApiImpl.DownloadListenerCreator;
 import io.flutter.plugins.webviewflutter.DownloadListenerHostApiImpl.DownloadListenerImpl;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -31,7 +32,7 @@ public class DownloadListenerTest {
 
   @Before
   public void setUp() {
-    instanceManager = new InstanceManager();
+    instanceManager = InstanceManager.open(identifier -> {});
 
     final DownloadListenerCreator downloadListenerCreator =
         new DownloadListenerCreator() {
@@ -46,6 +47,11 @@ public class DownloadListenerTest {
     hostApiImpl =
         new DownloadListenerHostApiImpl(instanceManager, downloadListenerCreator, mockFlutterApi);
     hostApiImpl.create(0L);
+  }
+
+  @After
+  public void tearDown() {
+    instanceManager.close();
   }
 
   @Test