+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,3 +1,11 @@
+## NEXT
+
+* Fixes avoid_redundant_argument_values lint warnings and minor typos.
+
+## 2.3.8
+
+* Updates ExoPlayer to 2.18.0.
+
 ## 2.3.7
 
 * Bumps gradle version to 7.2.1.